Change return type to double

+++ src/main/java/Unit.java
@@ -3,7 +3,7 @@ abstract public class Unit {
     static final Unit FOOT = new UnitFoot();
     static final Unit YARD = new UnitYard();
 
-    public abstract int rateToInch();
+    public abstract double rateToInch();
 
     public double rateTo(Unit unit) {
         return 1.0 * this.rateToInch() / unit.rateToInch();
